+++ CronJob/DiscordHook.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Reflection;
 using BepInEx.Bootstrap;
 using HarmonyLib;
@@ -8,18 +9,50 @@ public static class DiscordHook
 {
   public const string GUID = "games.nwest.valheim.discordconnector";
   private static Assembly? assembly;
-  public static void SendMessage(string message)
+  private static MethodInfo? sendMessage;
+  private static Type? eventType;
+  private static Dictionary<string, object> messageTypes = [];
+  public static void SendMessage(string type, string message)
   {
-    if (assembly == null) return;
-    var type = assembly.GetType("DiscordConnector.DiscordApi");
-    if (type == null) return;
-    var method = AccessTools.Method(type, "SendMessage", new Type[] { typeof(string) });
-    if (method == null) return;
-    method.Invoke(null, new object[] { message });
+    if (sendMessage == null || eventType == null) return;
+    if (!messageTypes.ContainsKey(type))
+    {
+      try
+      {
+        var messageType = Enum.Parse(eventType, type, true);
+        messageTypes[type] = messageType;
+      }
+      catch
+      {
+        CronJob.Log.LogError($"Invalid message type {type}");
+        return;
+      }
+    }
+    sendMessage.Invoke(null, [messageTypes[type], message]);
   }
   public static void Init()
   {
     if (!Chainloader.PluginInfos.TryGetValue(GUID, out var info)) return;
     assembly = info.Instance.GetType().Assembly;
+
+    eventType = assembly.GetType("DiscordConnector.Webhook+Event");
+    if (eventType == null)
+    {
+      CronJob.Log.LogError("Failed to get DiscordConnector.WebHook+Event");
+      return;
+    }
+    var type = assembly.GetType("DiscordConnector.DiscordApi");
+    if (type == null)
+    {
+      CronJob.Log.LogError("Failed to get DiscordConnector.DiscordApi");
+      return;
+    }
+    sendMessage = AccessTools.Method(type, "SendMessage", [eventType, typeof(string)]);
+    if (sendMessage == null)
+    {
+      CronJob.Log.LogError("Failed to get DiscordConnector.DiscordApi.SendMessage");
+      return;
+    }
+    CronJob.Log.LogInfo("DiscordConnector initialized");
   }
 }